+++ CarRental.Contracts/Cars/CarsCreateDto.cs
@@ -1,0 +1,14 @@
+﻿using System.ComponentModel.DataAnnotations;
+namespace CarRental.Application.Contracts.Cars;
+
+/// <summary>
+/// DTO for creating Car
+/// </summary>
+/// <param name="GenerationId">Generation</param>
+/// <param name="LicensePlate">License plate</param>
+/// <param name="Colour">Car colour</param>
+public record CarsCreateDto(
+    int GenerationId, 
+    string LicensePlate, 
+    string Colour
+);
