+++ CarRental.Contracts/Cars/CarsDto.cs
@@ -1,0 +1,15 @@
+﻿namespace CarRental.Application.Contracts.Cars;
+
+/// <summary>
+/// DTO for get requests to Cars
+/// </summary>
+/// <param name="Id">Car id</param>
+/// <param name="GenerationId">Generation</param>
+/// <param name="LicensePlate">License plate</param>
+/// <param name="Colour">Car colour</param>
+public record CarsDto(
+    int Id,
+    int GenerationId,
+    string LicensePlate,
+    string Colour
+);
