+++ CarRental.Contracts/RentalCars/RentalCarsCreateDto.cs
@@ -1,0 +1,15 @@
+﻿namespace CarRental.Application.Contracts.RentalCars;
+
+/// <summary>
+/// DTO for Rental car
+/// </summary>
+/// <param name="IssueTime">The date and time when the rental was issued/started</param>
+/// <param name="RentalHours">The duration of the rental in hours</param>
+/// <param name="RentedCarId">Renter car id</param>
+/// <param name="ClientId">Client id who rentering car</param>
+public record RentalCarsCreateDto(
+    DateTime IssueTime,
+    int RentalHours,
+    int RentedCarId,
+    int ClientId
+);
